LayoutTests/crypto/subtle/ec-import-jwk-key-export-raw-key-expected.txt

Test importing a jwk ECDH key and then export it in raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS bytesToHexString(publicKey) is expectedRawKey
PASS successfullyParsed is true

TEST COMPLETE
